… pyerrors/input/utils.py
… setup.py
… tests/sfcf_in_test.py