Hey there,

so as I mentioned in a previous pull request, I wanted to rewrite the read_sfcf method again.
So... I did. There is now a method "read_sfcf_multi", which is a superset of the "read_sfcf" method.
The read_sfcf method is now simply routed through the new method.
All of this was done for the following reason: I have tried to accelerate the reading of sfcf measurements in my projects for some time.
Finally, I found that the main time is spent in IO, as per read every file is opened once.
As it is usual to not read one correlator at a time, but e.g. 8 correlators at a time, this results in 8 $\times$ \#cnfg `open(file)` calls, which needs to retrieve the file from disk and loads it in the buffer.
Instead, now I am doing the following:
I directly read *all* relevant correlators from a file, then move on to the next file. Therefore I only need \#cnfg `open(file)` calls. The drawback being that afterwards I have to manage unpacking a python dict, which can be done much faster.

Here some benchmarks taken with `timeit`:

\#cnfg 2146

Old implementation:
\#corrs 72
read_sfcf 70.942 s

New implementation:
\#corrs 72
read_sfcf 71.462 s
read_sfcf_multi 4.777 s

\#corrs 36
read_sfcf 34.308 s
read_sfcf_multi 2.280 s

\#corrs 20
read_sfcf 18.421 s
read_sfcf_multi 1.653 s

This new functionality is not yet tested extensively (mainly I have done the tests by pytest that are also in this PR and the benchmarks on a set of measurements that I have for a project of mine).