@jkuhl-uni 2023-10-06
Hi,  mmmm... for now I have spent  my time oraganising the loops, as this is basically aall that is done here.
Maybe there is a way to more efficiently iterate through multiple levels of dicts that I am not aware of?
That would help a lot.