@jkuhl-uni 2023-10-16
Ah sorry, of course I can. I structured the returned dictionary in the following way:
`dict[name][quarks][offset][wf][wf2]`.
Of course, one could reorder the keys, but in any case I'd like the dict to have a nested structure, in which the user can search for what is needed.
My plan was furthermore, to "clean up" this dictionary, such that, e.g. if `len(name_list) == 1` the key is not needed after reading. With `nice_output = False`, one would have the complete nested structure, even if one just reads one single correlator.