@jkuhl-uni 2023-10-20
Hi,
thank you for your opinions, I couldn't agree more.
There are two things I could offer to control the code better:
1. more specialised tests
2. reintroduce templating in the code. I did this when using benedict and I think It could contribute to cleaner code. What I mean is, that very similar dict-structures are currently being generated at multiple places in the code. There should be a way around this. Although the complexity would stay the same, the code would become easier to read.
Maybe addressing these two things could benefit the code?