@fjosw 2023-10-20
Sounds good, your call if it's worth investing the time to rewrite stuff!