@jkuhl-uni 2023-10-20
Hi, so finally I refactored everything as Fabian suggested: In an earlier comment. as the result is put in its final norm in the last few lines of the method, I use the "flat" structure all the way through until there. I then added the option to either use nested dicts *or* a flat dict as the result.